Finished implementing instruction set

+++ asml/src/compiler/linker.rs
@@ -7,7 +7,7 @@ pub fn link(program: &mut Program) -> Result<(), String> {
             let label = label_val.1;
 
             if let Some(memloc) = program.labels.get(&label.label) {
-                let newloc = memloc + label.offset;
+                let newloc = memloc + label.offset as u16;
 
                 part.bytes[loc] = (newloc >> 8) as u8;
                 part.bytes[loc + 1] = newloc as u8;
